feat: refactor camera handling in HomeScreen and PinCodeBottomSheet, add cleanup in SnapshotManager

+++ app/src/main/java/net/invictusmanagement/invictuskiosk/presentation/qr_code_scanner/QRScannerViewModel.kt
@@ -109,7 +109,7 @@ class QRScannerViewModel @Inject constructor(
         lifecycleOwner: LifecycleOwner,
         previewView: PreviewView,
         executor: Executor,
-        currentAccessPointId: Int,
+        currentAccessPointId: Long,
     ) {
         val cameraProviderFuture = ProcessCameraProvider.getInstance(context)
